refactor: all integration table component spitted
